+++ api/endpoints/marca.py
@@ -3,17 +3,22 @@ from fastapi import APIRouter, status, Depends, HTTPException, Response
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.future import select
 from models.marca_model import MarcaModel
+from models.usuario_model import UsuarioModel
 from schemas.marca_schema import MarcaSchemaBase, MarcaSchema, MarcaSchemaUp
-from core.deps import get_session
+from core.deps import get_session, get_current_user
 
 
 router = APIRouter()
 
 
 # POST Marca
-@router.post('/', response_model=MarcaSchemaBase, status_code=status.HTTP_201_CREATED)
-async def post_marca(marca: MarcaSchema, db: AsyncSession= Depends(get_session)):
+@router.post('/', response_model=MarcaSchema, status_code=status.HTTP_201_CREATED)
+async def post_marca(marca: MarcaSchemaBase, usuario_logado: UsuarioModel = Depends(get_current_user), db: AsyncSession= Depends(get_session)):
     nova_marca: MarcaModel = MarcaModel(nome=marca.nome)
+
+    if usuario_logado.permissao_id != 1:
+        raise HTTPException(detail='O usuário logado não tem permissão para criar marcas.',
+                                status_code=status.HTTP_401_UNAUTHORIZED)
 
     db.add(nova_marca)
     await db.commit()
@@ -51,13 +56,19 @@ async def get_marca(marca_id: int, db: AsyncSession = Depends(get_session)):
                                 status_code=status.HTTP_404_NOT_FOUND)
         
 # PUT Marca
-@router.put('/{marca_id}', response_model=MarcaSchemaBase, status_code=status.HTTP_202_ACCEPTED)
-async def put_marca(marca_id: int, marca: MarcaSchemaUp, db: AsyncSession = Depends(get_session)):
+@router.put('/{marca_id}', response_model=MarcaSchema, status_code=status.HTTP_202_ACCEPTED)
+async def put_marca(marca_id: int, marca: MarcaSchemaUp, usuario_logado: UsuarioModel = Depends(get_current_user), 
+                    db: AsyncSession = Depends(get_session)):
+    
     async with db as session:
         query = select(MarcaModel).filter(MarcaModel.id == marca_id)
         result = await session.execute(query)
         marca_up: MarcaModel = result.scalars().unique().one_or_none()
 
+        if usuario_logado.permissao_id != 1:
+            raise HTTPException(detail='O usuário logado não tem permissão para editar marcas.',
+                                    status_code=status.HTTP_401_UNAUTHORIZED)
+        
         if marca_up:
             if marca.nome:
                 marca_up.nome = marca.nome
@@ -72,12 +83,16 @@ async def put_marca(marca_id: int, marca: MarcaSchemaUp, db: AsyncSession = Depe
 
 # DELETE Marca
 @router.delete('/{marca_id}', status_code=status.HTTP_204_NO_CONTENT)
-async def delete_produto(marca_id: int, db: AsyncSession = Depends(get_session)):
+async def delete_produto(marca_id: int, usuario_logado: UsuarioModel = Depends(get_current_user), db: AsyncSession = Depends(get_session)):
     async with db as session:
         query = select(MarcaModel).filter(MarcaModel.id == marca_id)
         result = await session.execute(query)
         marca_del: MarcaModel = result.scalars().unique().one_or_none()
 
+        if usuario_logado.permissao_id != 1:
+            raise HTTPException(detail='O usuário logado não tem permissão para deletar marcas.',
+                                    status_code=status.HTTP_401_UNAUTHORIZED)
+        
         if marca_del:
             await session.delete(marca_del)
             await session.commit()
