+++ api/endpoints/produto.py
@@ -4,20 +4,25 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.future import select
 from models.produto_model import ProdutoModel
+from models.usuario_model import UsuarioModel
 from schemas.produto_schema import ProdutoSchemaBase, ProdutoSchema, ProdutoSchemaUp
-from core.deps import get_session
+from core.deps import get_session, get_current_user
 
 
 router = APIRouter()
 
 
 # POST Produto
-@router.post('/', response_model=ProdutoSchemaBase, status_code=status.HTTP_201_CREATED)
-async def post_produto(produto: ProdutoSchemaBase, db: AsyncSession= Depends(get_session)):
+@router.post('/', response_model=ProdutoSchema, status_code=status.HTTP_201_CREATED)
+async def post_produto(produto: ProdutoSchemaBase, usuario_logado: UsuarioModel = Depends(get_current_user), db: AsyncSession= Depends(get_session)):
     novo_produto: ProdutoModel = ProdutoModel(nome=produto.nome, nome_referencia=produto.nome_referencia, 
                                               codigo=produto.codigo, marca_id=produto.marca_id, categoria_id=produto.categoria_id, 
                                               unidade_de_venda=produto.unidade_de_venda, valor=produto.valor )
     
+    if usuario_logado.permissao_id != 1:
+        raise HTTPException(detail='O usuário logado não tem permissão para criar produtos.',
+                                status_code=status.HTTP_401_UNAUTHORIZED)
+
     async with db as session:
         try:
             session.add(novo_produto)
@@ -26,7 +31,7 @@ async def post_produto(produto: ProdutoSchemaBase, db: AsyncSession= Depends(get
             return novo_produto
         except IntegrityError:
             raise HTTPException(status_code=status.HTTP_406_NOT_ACCEPTABLE,
-                                detail='Os dados inseridos são inválidos.')
+                                detail='Já existe um produto cadastrado com este código.')
 
 
 # GET Produtos
@@ -60,12 +65,16 @@ async def get_produto(produto_id: int, db: AsyncSession = Depends(get_session)):
 
 
 # PUT Produto
-@router.put('/{produto_id}', response_model=ProdutoSchemaBase, status_code=status.HTTP_202_ACCEPTED)
-async def put_produto(produto_id: int, produto: ProdutoSchemaUp, db: AsyncSession = Depends(get_session)):
+@router.put('/{produto_id}', response_model=ProdutoSchema, status_code=status.HTTP_202_ACCEPTED)
+async def put_produto(produto_id: int, produto: ProdutoSchemaUp, usuario_logado: UsuarioModel = Depends(get_current_user), db: AsyncSession = Depends(get_session)):
     async with db as session:
         query = select(ProdutoModel).filter(ProdutoModel.id == produto_id)
         result = await session.execute(query)
         produto_up: ProdutoModel = result.scalars().unique().one_or_none()
+
+        if usuario_logado.permissao_id != 1:
+            raise HTTPException(detail='O usuário logado não tem permissão para editar produtos.',
+                                    status_code=status.HTTP_401_UNAUTHORIZED)
 
         if produto_up:
             if produto.nome:
@@ -97,12 +106,16 @@ async def put_produto(produto_id: int, produto: ProdutoSchemaUp, db: AsyncSessio
 
 # DELETE Produto
 @router.delete('/{produto_id}', status_code=status.HTTP_204_NO_CONTENT)
-async def delete_produto(produto_id: int, db: AsyncSession = Depends(get_session)):
+async def delete_produto(produto_id: int, usuario_logado: UsuarioModel = Depends(get_current_user), db: AsyncSession = Depends(get_session)):
     async with db as session:
         query = select(ProdutoModel).filter(ProdutoModel.id == produto_id)
         result = await session.execute(query)
         produto_del: ProdutoModel = result.scalars().unique().one_or_none()
 
+        if usuario_logado.permissao_id != 1:
+            raise HTTPException(detail='O usuário logado não tem permissão para deletar produtos.',
+                                    status_code=status.HTTP_401_UNAUTHORIZED)
+        
         if produto_del:
             await session.delete(produto_del)
             await session.commit()
